requirements.txt
tailer==0.4.1
dnspython==2.0.0
